Guard against HTTP 403 error code when calling Picasa API directly
The OAuth access token expires quickly (every hour) and we get an HTTP
403 error code when we attempt to call a Picasa API with an expired
access token. The BloggerService class provided by the Google API Client
library for .NET takes care of automatically refreshing the access token
when calling the Blogger API. This is not true of our Picasa
integration, which features direct HTTP requests to the Picasa API
endpoint, so I've added code to catch the HTTP 403 error code and force
refresh the access token.