Creating a FileStream over a directory doesn't work very well
The root cause here is that we accidentally made a breaking change when
we removed the spell checking logic. We changed the
CONTEXT_DICTIONARY_FILE constant from "context.tlx" to string.Empty.
That means in the WriteSpellingContextDictionary() and
ReadSpellingContextDictionary() methods, we were resolving a dictionary
path as Path.Combine(SpellingContextDirectory, string.Empty); and then
tried to create a FileStream on a directory. The fix here is simple,
just change the dictionary filename back to "context.tlx".